+++ src/ILCompiler.Compiler/src/Compiler/CompilationModuleGroup.cs
@@ -2,8 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+using System.Reflection.Metadata.Ecma335;
+
 using Internal.TypeSystem;
 using Internal.TypeSystem.Ecma;
+using Internal.IL.Stubs.StartupCode;
 
 namespace ILCompiler
 {
@@ -11,6 +15,11 @@ namespace ILCompiler
     {
         protected CompilerTypeSystemContext _typeSystemContext;
         protected ICompilationRootProvider _rootProvider;
+
+        public MethodDesc StartupCodeMain
+        {
+            get; private set;
+        }
 
         protected CompilationModuleGroup(CompilerTypeSystemContext typeSystemContext, ICompilationRootProvider rootProvider)
         {
@@ -28,7 +37,7 @@ namespace ILCompiler
                 var module = _typeSystemContext.GetModuleFromPath(inputFile.Value);
 
                 if (module.PEReader.PEHeaders.IsExe)
-                    _rootProvider.AddMainMethodCompilationRoot(module);
+                    AddMainMethodCompilationRoot(module);
 
                 AddCompilationRootsForRuntimeExports(module);
             }
@@ -40,7 +49,7 @@ namespace ILCompiler
 
             if (IsTypeInCompilationGroup(stringType))
             {
-                _rootProvider.AddTypeCompilationRoot(stringType, "String type is always generated");
+                _rootProvider.AddCompilationRoot(stringType, "String type is always generated");
             }
         }
 
@@ -53,10 +62,24 @@ namespace ILCompiler
                     if (method.HasCustomAttribute("System.Runtime", "RuntimeExportAttribute"))
                     {
                         string exportName = ((EcmaMethod)method).GetAttributeStringValue("System.Runtime", "RuntimeExportAttribute");
-                        _rootProvider.AddMethodCompilationRoot(method, "Runtime export", exportName);
+                        _rootProvider.AddCompilationRoot(method, "Runtime export", exportName);
                     }
                 }
             }
+        }
+
+        private void AddMainMethodCompilationRoot(EcmaModule module)
+        {
+            if (StartupCodeMain != null)
+                throw new Exception("Multiple entrypoint modules");
+
+            int entryPointToken = module.PEReader.PEHeaders.CorHeader.EntryPointTokenOrRelativeVirtualAddress;
+            MethodDesc mainMethod = module.GetMethod(MetadataTokens.EntityHandle(entryPointToken));
+
+            var owningType = module.GetGlobalModuleType();
+            StartupCodeMain = new StartupCodeMainMethod(owningType, mainMethod);
+
+            _rootProvider.AddCompilationRoot(StartupCodeMain, "Startup Code Main Method", "__managed__Main");
         }
     }
 }
